+++ packages/engine-multi/src/test/bad-worker.js
@@ -1,0 +1,2 @@
+// This worker will not respond to a handshake task because no listener has been set up
+const handshake = () => true;