+++ packages/engine-multi/test/api/call-worker.test.ts
@@ -1,33 +1,41 @@
 import test from 'ava';
 import path from 'node:path';
 import EventEmitter from 'node:events';
-import { Promise as WorkerPoolPromise } from 'workerpool';
+import { createMockLogger } from '@openfn/logger';
 
 import initWorkers from '../../src/api/call-worker';
 import { EngineAPI } from '../../src/types';
-import { PURGE } from '../../src/events';
 
-let api = new EventEmitter() as EngineAPI;
+let engine = new EventEmitter() as EngineAPI;
 
-const workerPath = path.resolve('src/test/worker-functions.js');
+const workerPath = path.resolve('dist/test/worker-functions.js');
+const logger = createMockLogger();
 
 test.before(() => {
-  initWorkers(api, workerPath);
+  const { callWorker, closeWorkers } = initWorkers(
+    workerPath,
+    {
+      maxWorkers: 1,
+    },
+    logger
+  );
+  engine.callWorker = callWorker;
+  engine.closeWorkers = closeWorkers;
 });
 
-test.after(() => api.closeWorkers());
+test.after(() => engine.closeWorkers());
 
 test.serial('initWorkers should add a callWorker function', (t) => {
-  t.assert(typeof api.callWorker === 'function');
+  t.assert(typeof engine.callWorker === 'function');
 });
 
 test.serial('callWorker should return the default result', async (t) => {
-  const result = await api.callWorker('test', []);
+  const result = await engine.callWorker('test', []);
   t.is(result, 42);
 });
 
 test.serial('callWorker should return a custom result', async (t) => {
-  const result = await api.callWorker('test', [84]);
+  const result = await engine.callWorker('test', [84]);
   t.is(result, 84);
 });
 
@@ -38,85 +46,126 @@ test.serial('callWorker should trigger an event callback', async (t) => {
       done();
     };
 
-    api.callWorker('test', [11], { message: onCallback });
+    engine.callWorker('test', [11], { 'test-message': onCallback });
   });
 });
 
-test.serial(
+// Skipping because this routinely fails in CI
+test.serial.skip(
   'callWorker should throw TimeoutError if it times out',
   async (t) => {
-    await t.throwsAsync(() => api.callWorker('timeout', [11], {}, 10), {
-      instanceOf: WorkerPoolPromise.TimeoutError,
-    });
+    await t.throwsAsync(
+      () => engine.callWorker('wait', [5000], {}, { timeout: 10 }),
+      {
+        name: 'TimeoutError',
+      }
+    );
   }
 );
 
-// Dang, this doesn't work, the worker threads run in the same process
-test.skip('callWorker should execute with a different process id', async (t) => {
-  return new Promise((done) => {
-    const onCallback = ({ pid }) => {
-      t.not(process.pid, pid);
-      done();
+test.serial(
+  'callWorker should not freak out after a timeout error',
+  async (t) => {
+    await t.throwsAsync(
+      () => engine.callWorker('timeout', [11], {}, { timeout: 10 }),
+      {
+        name: 'TimeoutError',
+      }
+    );
+
+    const onCallback = (evt) => {
+      t.log(evt);
+      t.pass('all ok');
     };
 
-    api.callWorker('test', [], { message: onCallback });
+    await engine.callWorker('test', [], { 'test-message': onCallback });
+  }
+);
+
+test.serial('callWorker should execute in one process', async (t) => {
+  const ids: number[] = [];
+
+  await engine.callWorker('test', [], {
+    'test-message': ({ processId }) => {
+      ids.push(processId);
+    },
   });
+
+  await engine.callWorker('test', [], {
+    'test-message': ({ processId }) => {
+      ids.push(processId);
+    },
+  });
+
+  t.log(ids);
+  t.is(ids[0], ids[1]);
 });
 
-test.serial('callWorker should execute in a different process', async (t) => {
-  return new Promise((done) => {
-    // @ts-ignore
-    process.scribble = 'xyz';
+test.serial('callWorker should execute in two different threads', async (t) => {
+  const ids: number[] = [];
 
-    const onCallback = ({ scribble }) => {
-      // @ts-ignore
-      t.not(process.scribble, scribble);
-      done();
-    };
-
-    api.callWorker('test', [], { message: onCallback });
+  await engine.callWorker('test', [], {
+    'test-message': ({ threadId }) => {
+      ids.push(threadId);
+    },
   });
+
+  await engine.callWorker('test', [], {
+    'test-message': ({ threadId }) => {
+      ids.push(threadId);
+    },
+  });
+
+  t.log(ids);
+  t.not(ids[0], ids[1]);
 });
 
 test.serial(
-  'If null env is passed, worker thread should be able to access parent env',
+  'Even if null env is passed, worker thread should not be able to access parent env',
   async (t) => {
-    const badAPI = {} as EngineAPI;
     const env = null;
-    initWorkers(badAPI, workerPath, { env });
+    const { callWorker, closeWorkers } = initWorkers(
+      workerPath,
+      {
+        env,
+      },
+      logger
+    );
 
     // Set up a special key on process.env
     const code = '76ytghjs';
     process.env.TEST = code;
 
     // try and read that key inside the thread
-    const result = await badAPI.callWorker('readEnv', ['TEST']);
+    const result = await callWorker('readEnv', ['TEST']);
 
-    // voila, the kingdom is yours
-    t.is(result, code);
+    // Sorry pal, no dice
+    t.not(result, code);
 
-    badAPI.closeWorkers();
+    closeWorkers();
   }
 );
 
 test.serial(
   'By default, worker thread cannot access parent env if env not set (no options arg)',
   async (t) => {
-    const defaultAPI = {} as EngineAPI;
-
-    initWorkers(defaultAPI, workerPath /* no options passed*/);
+    const { callWorker, closeWorkers } = initWorkers(
+      workerPath,
+      undefined,
+      logger
+    );
 
     // Set up a special key on process.env
     const code = '76ytghjs';
     process.env.TEST = code;
 
     // try and read that key inside the thread
-    const result = await defaultAPI.callWorker('readEnv', ['TEST']);
+    const result = await callWorker('readEnv', ['TEST']);
 
     // No fish
     t.is(result, undefined);
 
-    defaultAPI.closeWorkers();
+    closeWorkers();
   }
 );
 
@@ -125,42 +174,49 @@ test.serial(
   async (t) => {
     const defaultAPI = {} as EngineAPI;
 
-    initWorkers(defaultAPI, workerPath, { maxWorkers: 1 });
+    const { callWorker, closeWorkers } = initWorkers(
+      workerPath,
+      { maxWorkers: 1 },
+      logger
+    );
 
     // Set up a special key on process.env
     const code = '76ytghjs';
     process.env.TEST = code;
 
     // try and read that key inside the thread
-    const result = await defaultAPI.callWorker('readEnv', ['TEST']);
+    const result = await callWorker('readEnv', ['TEST']);
 
     // No fish
     t.is(result, undefined);
 
-    defaultAPI.closeWorkers();
+    closeWorkers();
   }
 );
 
 test.serial(
   'Worker thread cannot access parent env if custom env is passted',
   async (t) => {
-    const customAPI = {} as EngineAPI;
     const env = { NODE_ENV: 'production' };
-    initWorkers(customAPI, workerPath, { env });
+    const { callWorker, closeWorkers } = initWorkers(
+      workerPath,
+      { env },
+      logger
+    );
 
     // Set up a special key on process.env
     const code = '76ytghjs';
     process.env.TEST = code;
 
     // try and read that key inside the thread
-    const result = await customAPI.callWorker('readEnv', ['TEST']);
+    const result = await callWorker('readEnv', ['TEST']);
 
     // No fish
     t.is(result, undefined);
 
-    const result2 = await customAPI.callWorker('readEnv', ['NODE_ENV']);
+    const result2 = await callWorker('readEnv', ['NODE_ENV']);
     t.is(result2, 'production');
 
-    customAPI.closeWorkers();
+    closeWorkers();
   }
 );