+++ packages/engine-multi/tsup.config.js
@@ -5,8 +5,10 @@ export default {
   splitting: false,
   entry: {
     index: 'src/index.ts',
-    'worker/worker': 'src/worker/worker.ts',
-    // TODO I don't actually want to build this into the dist, I don't think?
-    'worker/mock': 'src/worker/mock-worker.ts',
+    'worker/thread/run': 'src/worker/thread/run.ts',
+    'worker/child/runner': 'src/worker/child/runner.ts',
+
+    'test/mock-run': 'src/worker/thread/mock-run.ts',
+    'test/worker-functions': 'src/test/worker-functions.ts',
   },
 };